use generic (#20)
* use generic

* add boost-beast to build

* update 123

* add login client wip

* use login server

* disable login for now

* add get_ticks_ms func

* add json parse

* fix crash

* add disable login for now.

* add disable login

* disable login

* update123

* fix crash

* update123

* update123

* fix timestamp

* add on reload method when reloading server

* update123

+++ src/any_type.h
@@ -100,7 +100,7 @@ static bool operator!=(const AnyElement &lhs, const AnyElement &rhs) { return !(
 
 std::string decode_string_or_default(yyjson_val *object, std::string key,
                                      std::string default_value);
-int decode_int_or_default(yyjson_val *object, std::string key, int default_value);
+int64_t decode_int_or_default(yyjson_val *object, std::string key, int64_t default_value);
 std::string decode_array(yyjson_val *array, AnyElement &element);
 std::string decode_value(yyjson_val *value, AnyElement &element);
 std::string decode_object(yyjson_val *object,
